Fix some warning catched by clang in 64bit.

+++ Quicksilver/Code-QuickStepEffects/QSShading.m
@@ -23,9 +23,9 @@ struct QSGradientPoint {
 }
 */
 
-void QSColorFade(void *info, const float *in, float *out) {
-	float v = *in;
-	float *colors = info;
+void QSColorFade(void *info, const CGFloat *in, CGFloat *out) {
+	CGFloat v = *in;
+	CGFloat *colors = info;
 	int i;
 	for (i = 0; i<4; i++)
 		*out++ = colors[i] *(1-v) +colors[i+4] *(v);
@@ -36,8 +36,8 @@ void QSFillRectWithGradientFromEdge(NSRect rect, NSColor *start, NSColor *end, N
 	start = [start colorUsingColorSpaceName:NSDeviceRGBColorSpace];
 	end = [end colorUsingColorSpaceName:NSDeviceRGBColorSpace];
 	size_t components;
-	static const float domain[2] = { 0, 1 } ;
-	static const float range[10] = { 0, 1, 0, 1, 0, 1, 0, 1, 0, 1 } ;
+	static const CGFloat domain[2] = { 0, 1 } ;
+	static const CGFloat range[10] = { 0, 1, 0, 1, 0, 1, 0, 1, 0, 1 } ;
 	static const CGFunctionCallbacks callbacks = { 0, &QSColorFade, NULL } ;
 
 	float colors[8] = {
